pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scientia"
version = "0.1.0"
description = "Generate verified ScienceClaw-compatible skills from any scientific source"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "requests>=2.31.0",
    "click>=8.1.0",
    "pypdf>=4.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
browser = ["playwright>=1.40.0"]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "responses>=0.25.0"]

[project.scripts]
scientia = "scientia.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.setuptools.packages.find]
where = ["."]
include = ["scientia*"]

[tool.setuptools.package-dir]
"" = "."
